feat:外部ui自动化脚本上传;编辑;删除;列表

+++ app/api/ui_auto_file_api/ui_auto_file_api.py
@@ -5,12 +5,32 @@
 # @File    : ui_auto_file_api.py
 # @Software: PyCharm
 
+from flask import current_app
+
 from all_reference import *
-from config.config import get_config
 from app.models.ui_auto_file.models import UiAutoFile
 
-conf = get_config()
-UPLOAD_PATH = conf.get('base', 'UPLOAD_PATH')
+
+def check_files(path, new_folder, new_files):
+    for root, folder, file in os.walk(path):
+        if root == path:
+            print(f"root:{root}")
+            print(f"folder:{folder}")
+            print(f"file:{file}")
+            if new_folder not in folder:
+                os.mkdir(f'{root}/{new_folder}')
+                return True, 'pass'
+            else:
+                user_dir = list(os.walk(f"{root}/{new_folder}"))[0]
+                print(user_dir)
+                user_files = user_dir[2]
+                print(user_files)
+                for index, f in enumerate(new_files):
+                    if f in user_files:
+                        return False, f'文件: {f} 已经存在'
+                return True, 'pass'
+        else:
+            return False, f"路径: {root} 错误"
 
 
 class UiAutoFileApi(MethodView):
@@ -24,55 +44,86 @@ class UiAutoFileApi(MethodView):
         query_file = UiAutoFile.query.get(file_id)
 
         if not query_file:
-            return api_result(code=400, message='文件id:{}不存在'.format(file_id))
+            return api_result(code=400, message=f'文件id:{file_id}不存在')
 
         return api_result(code=200, message='操作成功', data=query_file.to_json())
 
     def post(self):
         """脚本上传"""
 
-        file_list = [f for f in os.listdir(UPLOAD_PATH)]
-        file = request.files.get('file')
+        # file = request.files.get('file')  # 单个文件
+        file_list = request.files.getlist('file')  # 多个文件
         title = request.form.get('title')
         remark = request.form.get('remark')
+        upload_path = current_app.config.get("UPLOAD_PATH")
 
-        file_name = file.filename
+        if not file_list:
+            return api_result(code=400, message='文件不能为空')
 
-        if file_name in file_list:
-            return api_result(code=400, message=f'文件名称已存在:{filename}')
+        if len(file_list) > 9:
+            return api_result(code=400, message='单次文件上传不能多于9个')
 
-        file.save(f'{UPLOAD_PATH}/{file_name}')
+        new_files = [f.filename for f in file_list]
+        print(new_files)
+        _bool, _message = check_files(path=upload_path, new_folder=g.app_user.username, new_files=new_files)
+        print(_bool, _message)
 
-        new_upload = UiAutoFile(
-            title=title,
-            file_name=file_name,
-            file_path=UPLOAD_PATH,
-            remark=remark,
-            creator=g.app_user.username,
-            creator_id=g.app_user.id,
-        )
-        new_upload.save()
+        if not _bool:
+            return api_result(code=400, message=f'{_message}')
+
+        for file in file_list:
+            file_name = file.filename
+            save_path = f'{upload_path}/{g.app_user.username}/{file_name}'
+            file.save(save_path)
+            new_upload = UiAutoFile(
+                title=title,
+                file_name=file_name,
+                file_path=save_path,
+                remark=remark,
+                creator=g.app_user.username,
+                creator_id=g.app_user.id,
+            )
+            db.session.add(new_upload)
+        db.session.commit()
+
         return api_result(code=201, message='操作成功')
 
     def put(self):
         """编辑脚本"""
 
-        data = request.get_json()
-        file_id = data.get('id')
-        title = data.get('title')
-        remark = data.get('remark')
+        file = request.files.get('file')
+        file_id = request.form.get('id')
+        title = request.form.get('title')
+        remark = request.form.get('remark')
+        upload_path = current_app.config.get("UPLOAD_PATH")
+
+        if not file:
+            return api_result(code=400, message='文件不能为空')
 
         query_file = UiAutoFile.query.get(file_id)
 
         if not query_file:
-            return api_result(code=400, message='file_id:不存在')
+            return api_result(code=400, message=f'file_id: {file_id} 不存在')
+
+        if query_file.creator_id != g.app_user.id:
+            return api_result(code=400, message='只允许创建者编辑')
+
+        if file:
+            file_name = file.filename
+            save_path = f'{upload_path}/{g.app_user.username}/{file_name}'
+            query_file.file_name = file_name
+            query_file.file_path = save_path
+            try:
+                os.remove(query_file.file_path)
+                file.save(save_path)
+            except BaseException as e:
+                file.save(save_path)
 
         query_file.title = title
         query_file.remark = remark
         query_file.modifier = g.app_user.username
         query_file.modifier_id = g.app_user.id
         db.session.commit()
-
         return api_result(code=203, message='操作成功')
 
     def delete(self):
@@ -84,11 +135,15 @@ class UiAutoFileApi(MethodView):
         query_file = UiAutoFile.query.get(file_id)
 
         if not query_file:
-            return api_result(code=400, message='file_id:不存在')
+            return api_result(code=400, message=f'file_id: {file_id} 不存在')
+
+        if query_file.creator_id != g.app_user.id:
+            return api_result(code=400, message='只允许创建者删除')
 
         query_file.modifier = g.app_user.username
         query_file.modifier_id = g.app_user.id
         query_file.delete()
+        os.remove(query_file.file_path)
 
         return api_result(code=204, message='操作成功')
 
@@ -107,7 +162,6 @@ class UiAutoFilePageApi(MethodView):
         title = data.get('title')
         file_name = data.get('file_name')
         creator_id = data.get('creator_id')
-        is_deleted = data.get('is_deleted', 0)
         page = data.get('page')
         size = data.get('size')
 
@@ -125,8 +179,8 @@ class UiAutoFilePageApi(MethodView):
 
         where_dict = {
             "id": file_id,
-            "is_deleted": is_deleted,
-            "creator_id": creator_id,
+            "is_deleted": 0,
+            "creator_id": creator_id
         }
 
         result_data = general_query(
@@ -141,7 +195,7 @@ class UiAutoFilePageApi(MethodView):
         return api_result(code=200, message='操作成功', data=result_data)
 
 
-class UiAutoFileCallApi(MethodView):
+class UiAutoFileCallAioApi(MethodView):
     """
     UI自动化测试脚本文件调用执行api
     """
